requirements.txt
aniso8601==9.0.1
APIFlask==1.3.0
apispec==6.3.0
APScheduler==3.10.1
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
filelock==3.10.0
Flask==2.2.3
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-Script==2.0.5
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
lit==15.0.7
MarkupSafe==2.1.2
marshmallow==3.19.0
mistune==2.0.5
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.1
triton==2.0.0
typing-extensions==3.10.0.2
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
webargs==8.2.0
Werkzeug==2.2.3
zipp==3.15.0
